Function web_server::config::settings::get_local_time

pub fn get_local_time(
    now: DateTime<Utc>,
) -> Result<DateTime<FixedOffset>, Box<dyn Error>>

Converts a UTC datetime to the local time based on the configured timezone offset.

Functionality

This function performs the following steps:

Retrieves the timezone offset using the get_timezone_offset() function.

Applies this offset to the provided UTC datetime to convert it to local time.

Parameters

now: A DateTime<Utc> representing the UTC time to be converted.

Returns

Returns a Result<DateTime<FixedOffset>, Box<dyn Error>>:

Ok(DateTime<FixedOffset>): The local time as a DateTime with a FixedOffset if successful.

Err(Box<dyn Error>): An error if the timezone offset couldn’t be retrieved.

Errors

This function will return an error if get_timezone_offset() fails, which can happen if:

The TIMEZONE_OFFSET environment variable is not set or invalid.

The resulting offset is invalid.

Example

let utc_now = Utc::now();
match get_local_time(utc_now) {
    Ok(local_time) => println!("Local time: {}", local_time),
    Err(e) => eprintln!("Error converting to local time: {}", e),
}

Note

This function relies on the get_timezone_offset() function, so ensure that the TIMEZONE_OFFSET is correctly set in your environment before use.